Added IndexFunctor concept.

+++ include/satyr/tbb_for.h
@@ -12,7 +12,8 @@ template <Policy Policy, class F>
 void for_(Policy policy, index_t first, index_t last, F f) {
   auto grainularity = get_policy<satyr::grainularity>(policy);
   tbb::parallel_for(
-      tbb::blocked_range<index_t>{first, last, grainularity.value},
+      tbb::blocked_range<index_t>{first, last,
+                                  static_cast<size_t>(grainularity.value)},
       [=](const tbb::blocked_range<index_t>& range) {
         for_(policy | serial_v, range.begin(), range.end(), f);
       });
